requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
PySocks==1.7.0
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
slackclient==1.3.1
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.2
